ref(emitter): split CallEmitter god-class into per-family collaborators
Extract the 23 inline call-site specialization emission methods from the
959-LOC CallEmitter into eight focused Specialized/*CallEmitter classes,
one per eligibility family (TypedValue, CoreFn, NilBoolean, TypePredicate,
AtomMethod, TypedCollection, AssocConj, NumericOperation), each implementing
SpecializedCallEmitterInterface::tryEmit(CallNode): bool.

CallEmitter now builds the collaborators once in its constructor (cached per
OutputEmitter by NodeEmitterFactory) and dispatches by looping tryEmit() over
them before the generic call path, dropping to a thin ~310-LOC dispatcher.
Family predicates are disjoint, so emitted PHP is byte-for-byte unchanged;
verified by the integration fixture suite and a perf-neutral compile bench.
