+++ src/php/Compiler/Domain/Emitter/OutputEmitter/NodeEmitter/CallEmitter.php
@@ -11,27 +11,49 @@ use Phel\Compiler\Domain\Analyzer\Ast\GlobalVarNode;
 use Phel\Compiler\Domain\Analyzer\Ast\LocalVarNode;
 use Phel\Compiler\Domain\Analyzer\Ast\PhpClassNameNode;
 use Phel\Compiler\Domain\Analyzer\Ast\PhpVarNode;
-use Phel\Compiler\Domain\Emitter\OutputEmitter\AssocConjSpecialization;
-use Phel\Compiler\Domain\Emitter\OutputEmitter\AtomMethodSpecialization;
 use Phel\Compiler\Domain\Emitter\OutputEmitter\CallSpecialization;
 use Phel\Compiler\Domain\Emitter\OutputEmitter\GlobalCallTarget;
-use Phel\Compiler\Domain\Emitter\OutputEmitter\NilAndBooleanCheckSpecialization;
+use Phel\Compiler\Domain\Emitter\OutputEmitter\NodeEmitter\Specialized\AssocConjCallEmitter;
+use Phel\Compiler\Domain\Emitter\OutputEmitter\NodeEmitter\Specialized\AtomMethodCallEmitter;
+use Phel\Compiler\Domain\Emitter\OutputEmitter\NodeEmitter\Specialized\CoreFnCallEmitter;
+use Phel\Compiler\Domain\Emitter\OutputEmitter\NodeEmitter\Specialized\NilBooleanCallEmitter;
+use Phel\Compiler\Domain\Emitter\OutputEmitter\NodeEmitter\Specialized\NumericOperationCallEmitter;
+use Phel\Compiler\Domain\Emitter\OutputEmitter\NodeEmitter\Specialized\SpecializedCallEmitterInterface;
+use Phel\Compiler\Domain\Emitter\OutputEmitter\NodeEmitter\Specialized\TypedCollectionCallEmitter;
+use Phel\Compiler\Domain\Emitter\OutputEmitter\NodeEmitter\Specialized\TypedValueCallEmitter;
+use Phel\Compiler\Domain\Emitter\OutputEmitter\NodeEmitter\Specialized\TypePredicateCallEmitter;
 use Phel\Compiler\Domain\Emitter\OutputEmitter\NodeEmitterInterface;
-use Phel\Compiler\Domain\Emitter\OutputEmitter\NumericOperationSpecialization;
 use Phel\Compiler\Domain\Emitter\OutputEmitter\PhpStringEscape;
-use Phel\Compiler\Domain\Emitter\OutputEmitter\TypedCollectionMethodSpecialization;
-use Phel\Compiler\Domain\Emitter\OutputEmitter\TypedValueSpecialization;
-use Phel\Compiler\Domain\Emitter\OutputEmitter\TypePredicateSpecialization;
+use Phel\Compiler\Domain\Emitter\OutputEmitterInterface;
 use Phel\Lang\Symbol;
 
-use function array_slice;
 use function assert;
 use function count;
-use function sprintf;
 
-final class CallEmitter implements NodeEmitterInterface
+final readonly class CallEmitter implements NodeEmitterInterface
 {
-    use WithOutputEmitterTrait;
+    /** @var list<SpecializedCallEmitterInterface> */
+    private array $specializedEmitters;
+
+    public function __construct(
+        private OutputEmitterInterface $outputEmitter,
+    ) {
+        // Ordered chain of call-site specialisation families. Each family
+        // returns false when its eligibility predicates do not match, so the
+        // node falls through to the generic dispatch below. The predicates
+        // across families are disjoint (distinct fn names, arities, and
+        // analyser tags), so the order between families is not significant.
+        $this->specializedEmitters = [
+            new TypedValueCallEmitter($outputEmitter),
+            new CoreFnCallEmitter($outputEmitter),
+            new NilBooleanCallEmitter($outputEmitter),
+            new TypePredicateCallEmitter($outputEmitter),
+            new AtomMethodCallEmitter($outputEmitter),
+            new TypedCollectionCallEmitter($outputEmitter),
+            new AssocConjCallEmitter($outputEmitter),
+            new NumericOperationCallEmitter($outputEmitter),
+        ];
+    }
 
     public function emit(AbstractNode $node): void
     {
@@ -163,96 +185,10 @@ final class CallEmitter implements NodeEmitterInterface
             return;
         }
 
-        if ($this->tryEmitKeywordFind($node)) {
-            return;
-        }
-
-        if ($this->tryEmitStrConcat($node)) {
-            return;
-        }
-
-        if ($this->tryEmitTypedGetAccess($node)) {
-            return;
-        }
-
-        if ($this->tryEmitTypedPhpArrayGet($node)) {
-            return;
-        }
-
-        if ($this->tryEmitTypedPhpArrayCount($node)) {
-            return;
-        }
-
-        if ($this->tryEmitNilCheck($node)) {
-            return;
-        }
-
-        if ($this->tryEmitSomeCheck($node)) {
-            return;
-        }
-
-        if ($this->tryEmitTrueCheck($node)) {
-            return;
-        }
-
-        if ($this->tryEmitFalseCheck($node)) {
-            return;
-        }
-
-        if ($this->tryEmitTruthyCheck($node)) {
-            return;
-        }
-
-        if ($this->tryEmitNumericPredicate($node)) {
-            return;
-        }
-
-        if ($this->tryEmitTypePredicate($node)) {
-            return;
-        }
-
-        if ($this->tryEmitNamedAccessor($node)) {
-            return;
-        }
-
-        if ($this->tryEmitEmptyCheck($node)) {
-            return;
-        }
-
-        if ($this->tryEmitContainsCheck($node)) {
-            return;
-        }
-
-        if ($this->tryEmitAtomMethodCall($node)) {
-            return;
-        }
-
-        if ($this->tryEmitTypedVectorAccessor($node)) {
-            return;
-        }
-
-        if ($this->tryEmitTypedSeqAccessor($node)) {
-            return;
-        }
-
-        if ($this->tryEmitAssocConjChain($node)) {
-            return;
-        }
-
-        if ($this->tryEmitTypedAssocConjDissoc($node)) {
-            return;
-        }
-
-        if ($this->tryEmitNotEqPeephole($node)) {
-            return;
-        }
-
-        if ($this->tryEmitTypedVariadicChain($node)) {
-            return;
-        }
-
-        if ($this->tryEmitTypedBinaryOp($node)) {
-            return;
+        foreach ($this->specializedEmitters as $specializedEmitter) {
+            if ($specializedEmitter->tryEmit($node)) {
+                return;
+            }
         }
 
         $useCallMethod = !$this->isSelfCall($node)
@@ -265,588 +201,6 @@ final class CallEmitter implements NodeEmitterInterface
         } else {
             $this->emitFunctionArguments($node);
         }
-    }
-
-    /**
-     * Specialise `(str ...)` to PHP `.` concatenation when every arg
-     * compiles to a string-typed expression. The runtime `phel.core/str`
-     * does a per-arg `val-to-str` dispatch plus a `StringBuilder`-style
-     * accumulator pass; when every arg is already a string the result is
-     * the same plain `.` chain, so we emit it directly and skip both the
-     * registry lookup and the runtime walk.
-     *
-     * Eligibility lives on {@see CallSpecialization::isStrConcat()} so the
-     * cache scanner can skip reserving a `static $__phel_call_N` slot for
-     * the call we are about to specialise.
-     */
-    private function tryEmitStrConcat(CallNode $node): bool
-    {
-        if (!CallSpecialization::isStrConcat($node)) {
-            return false;
-        }
-
-        $loc = $node->getStartSourceLocation();
-        $this->outputEmitter->emitStr('(', $loc);
-        foreach ($node->getArguments() as $i => $arg) {
-            if ($i > 0) {
-                $this->outputEmitter->emitStr(' . ', $loc);
-            }
-
-            $this->outputEmitter->emitNode($arg);
-        }
-
-        $this->outputEmitter->emitStr(')', $loc);
-        return true;
-    }
-
-    /**
-     * Specialise `(:k m)` to `$m->find(\Phel::keyword("k"))` when the
-     * analyser has proved `m` to be a `PersistentMapInterface`. The
-     * runtime `Keyword::__invoke` dispatches on the target's runtime
-     * type to pick between `ArrayAccess`, `ContainsInterface`, and the
-     * `nil` fallback; a typed map collapses that dispatch to the single
-     * `find` call the map already exposes, returning `null` on miss to
-     * match the 1-arg keyword-accessor contract.
-     */
-    private function tryEmitKeywordFind(CallNode $node): bool
-    {
-        if (!TypedValueSpecialization::isKeywordFind($node)) {
-            return false;
-        }
-
-        $loc = $node->getStartSourceLocation();
-        $this->outputEmitter->emitStr('(', $loc);
-        $this->outputEmitter->emitNode($node->getArguments()[0]);
-        $this->outputEmitter->emitStr('->find(', $loc);
-        $this->outputEmitter->emitNode($node->getFn());
-        $this->outputEmitter->emitStr('))', $loc);
-        return true;
-    }
-
-    /**
-     * Specialise two-arg `phel.core` arithmetic / comparison wrappers
-     * to the native PHP binary op when both args are statically proven
-     * primitive. The runtime defns route through `NumericOperations`
-     * to handle `BigInt` / `Ratio` polymorphism; for primitive-typed
-     * call sites that dispatch is wasted work and collapses to a
-     * single PHP operator.
-     *
-     * Eligibility lives on {@see NumericOperationSpecialization::typedBinaryOpName()}.
-     */
-    private function tryEmitTypedBinaryOp(CallNode $node): bool
-    {
-        $op = NumericOperationSpecialization::typedBinaryOpName($node);
-        if ($op === null) {
-            return false;
-        }
-
-        $args = $node->getArguments();
-        $loc = $node->getStartSourceLocation();
-        $this->outputEmitter->emitStr('(', $loc);
-        $this->outputEmitter->emitNode($args[0]);
-        $this->outputEmitter->emitStr(' ' . $op . ' ', $loc);
-        $this->outputEmitter->emitNode($args[1]);
-        $this->outputEmitter->emitStr(')', $loc);
-        return true;
-    }
-
-    /**
-     * Variadic (N>=3) numeric / ordering ops over tagged numeric locals.
-     * `arith` ops chain as `($a + $b + $c)` (PHP is left-associative just
-     * like Phel's variadic `+`/`*`); `compare` ops expand to a pairwise
-     * `&&` chain `(($a < $b) && ($b < $c))` because PHP `<` does not
-     * thread its result through the next comparison the way Phel does.
-     *
-     * Eligibility lives on {@see NumericOperationSpecialization::typedVariadicChain()}.
-     */
-    private function tryEmitTypedVariadicChain(CallNode $node): bool
-    {
-        $spec = NumericOperationSpecialization::typedVariadicChain($node);
-        if ($spec === null) {
-            return false;
-        }
-
-        $args = $node->getArguments();
-        $loc = $node->getStartSourceLocation();
-        $op = $spec['op'];
-        $this->outputEmitter->emitStr('(', $loc);
-
-        if ($spec['kind'] === 'arith') {
-            $this->outputEmitter->emitNode($args[0]);
-            for ($i = 1, $n = count($args); $i < $n; ++$i) {
-                $this->outputEmitter->emitStr(' ' . $op . ' ', $loc);
-                $this->outputEmitter->emitNode($args[$i]);
-            }
-        } else {
-            for ($i = 0, $n = count($args) - 1; $i < $n; ++$i) {
-                if ($i > 0) {
-                    $this->outputEmitter->emitStr(' && ', $loc);
-                }
-
-                $this->outputEmitter->emitStr('(', $loc);
-                $this->outputEmitter->emitNode($args[$i]);
-                $this->outputEmitter->emitStr(' ' . $op . ' ', $loc);
-                $this->outputEmitter->emitNode($args[$i + 1]);
-                $this->outputEmitter->emitStr(')', $loc);
-            }
-        }
-
-        $this->outputEmitter->emitStr(')', $loc);
-        return true;
-    }
-
-    /**
-     * Specialise `(get coll k)` to a direct method call when the target
-     * carries a `PersistentVectorInterface` or `PersistentMapInterface`
-     * tag. Skips the cond chain in `phel.core/get`'s body (nil / set /
-     * seq / php-aget fallback) for the hot indexed-access shape.
-     *
-     * Two-arg form only: the three-arg `(get coll k default)` shape
-     * needs an explicit `contains?` probe to honour the default, so
-     * the cond chain is still the right path.
-     */
-    private function tryEmitTypedGetAccess(CallNode $node): bool
-    {
-        $method = CallSpecialization::typedGetAccessMethod($node);
-        if ($method === null) {
-            return false;
-        }
-
-        $args = $node->getArguments();
-        $loc = $node->getStartSourceLocation();
-        $this->outputEmitter->emitStr('(', $loc);
-        $this->outputEmitter->emitNode($args[0]);
-        $this->outputEmitter->emitStr('->' . $method . '(', $loc);
-        $this->outputEmitter->emitNode($args[1]);
-        $this->outputEmitter->emitStr('))', $loc);
-        return true;
-    }
-
-    /**
-     * Specialise `(get arr k)` / `(get arr k default)` on a target
-     * tagged `array` to a native PHP subscript with the null-coalescing
-     * fallback. Matches the runtime `get` semantics for PHP arrays —
-     * `(php/aget ds k)` then "if nil return default" — because PHP's
-     * `??` treats both absent keys and explicit nulls as triggering
-     * the fallback.
-     */
-    /**
-     * Specialise `(count arr)` on a target tagged `array` to a native
-     * `count(\$arr)` call. The runtime body would walk a cond chain
-     * over the standard collection shapes before reaching the same
-     * `php/count` branch.
-     */
-    private function tryEmitTypedPhpArrayCount(CallNode $node): bool
-    {
-        if (!CallSpecialization::isTypedPhpArrayCount($node)) {
-            return false;
-        }
-
-        $loc = $node->getStartSourceLocation();
-        $this->outputEmitter->emitStr('count(', $loc);
-        $this->outputEmitter->emitNode($node->getArguments()[0]);
-        $this->outputEmitter->emitStr(')', $loc);
-        return true;
-    }
-
-    /**
-     * `(nil? x)` — emit `($x === null)` directly, bypassing the
-     * registry lookup and `id` adapter.
-     */
-    private function tryEmitNilCheck(CallNode $node): bool
-    {
-        if (!NilAndBooleanCheckSpecialization::isNilCheck($node)) {
-            return false;
-        }
-
-        $loc = $node->getStartSourceLocation();
-        $this->outputEmitter->emitStr('(', $loc);
-        $this->outputEmitter->emitNode($node->getArguments()[0]);
-        $this->outputEmitter->emitStr(' === null)', $loc);
-        return true;
-    }
-
-    /**
-     * `(some? x)` 1-arg — emit `($x !== null)` directly. The 2-arg
-     * overload `(some? pred coll)` keeps the runtime dispatch.
-     */
-    private function tryEmitSomeCheck(CallNode $node): bool
-    {
-        if (!NilAndBooleanCheckSpecialization::isSomeCheck($node)) {
-            return false;
-        }
-
-        $loc = $node->getStartSourceLocation();
-        $this->outputEmitter->emitStr('(', $loc);
-        $this->outputEmitter->emitNode($node->getArguments()[0]);
-        $this->outputEmitter->emitStr(' !== null)', $loc);
-        return true;
-    }
-
-    private function tryEmitTrueCheck(CallNode $node): bool
-    {
-        if (!NilAndBooleanCheckSpecialization::isTrueCheck($node)) {
-            return false;
-        }
-
-        $loc = $node->getStartSourceLocation();
-        $this->outputEmitter->emitStr('(', $loc);
-        $this->outputEmitter->emitNode($node->getArguments()[0]);
-        $this->outputEmitter->emitStr(' === true)', $loc);
-        return true;
-    }
-
-    private function tryEmitFalseCheck(CallNode $node): bool
-    {
-        if (!NilAndBooleanCheckSpecialization::isFalseCheck($node)) {
-            return false;
-        }
-
-        $loc = $node->getStartSourceLocation();
-        $this->outputEmitter->emitStr('(', $loc);
-        $this->outputEmitter->emitNode($node->getArguments()[0]);
-        $this->outputEmitter->emitStr(' === false)', $loc);
-        return true;
-    }
-
-    /**
-     * `(truthy? x)` — Phel-truthy probe inlined. Uses a fresh `$__truthy`
-     * binding so the result is a bool the caller can splice into any
-     * expression position.
-     */
-    private function tryEmitTruthyCheck(CallNode $node): bool
-    {
-        if (!NilAndBooleanCheckSpecialization::isTruthyCheck($node)) {
-            return false;
-        }
-
-        $loc = $node->getStartSourceLocation();
-        $this->outputEmitter->emitStr('(($__truthy = ', $loc);
-        $this->outputEmitter->emitNode($node->getArguments()[0]);
-        $this->outputEmitter->emitStr(') !== null && $__truthy !== false)', $loc);
-        return true;
-    }
-
-    /**
-     * `(deref x)` / `(reset! v val)` — emit the direct method call on
-     * the target, skipping the registry lookup. No tag required: the
-     * runtime body itself is a single `php/-> target (method ...)`, so
-     * the failure mode (method not found) is identical to today.
-     */
-    private function tryEmitAtomMethodCall(CallNode $node): bool
-    {
-        $shape = AtomMethodSpecialization::atomMethodCall($node);
-        if ($shape === null) {
-            return false;
-        }
-
-        [$method, $argIndices] = $shape;
-        $args = $node->getArguments();
-        $loc = $node->getStartSourceLocation();
-
-        $this->outputEmitter->emitStr('(', $loc);
-        $this->outputEmitter->emitNode($args[0]);
-        $this->outputEmitter->emitStr('->' . $method . '(', $loc);
-        foreach ($argIndices as $i => $idx) {
-            if ($i > 0) {
-                $this->outputEmitter->emitStr(', ', $loc);
-            }
-
-            $this->outputEmitter->emitNode($args[$idx]);
-        }
-
-        $this->outputEmitter->emitStr('))', $loc);
-        return true;
-    }
-
-    /**
-     * `(contains? coll k)` on a tagged target — emit the direct
-     * method or `array_key_exists` form.
-     */
-    private function tryEmitContainsCheck(CallNode $node): bool
-    {
-        $kind = TypedValueSpecialization::containsCheckKind($node);
-        if ($kind === null) {
-            return false;
-        }
-
-        $args = $node->getArguments();
-        $loc = $node->getStartSourceLocation();
-
-        if ($kind === 'array') {
-            $this->outputEmitter->emitStr('array_key_exists(', $loc);
-            $this->outputEmitter->emitNode($args[1]);
-            $this->outputEmitter->emitStr(', ', $loc);
-            $this->outputEmitter->emitNode($args[0]);
-            $this->outputEmitter->emitStr(')', $loc);
-            return true;
-        }
-
-        $this->outputEmitter->emitStr('(', $loc);
-        $this->outputEmitter->emitNode($args[0]);
-        $this->outputEmitter->emitStr('->contains(', $loc);
-        $this->outputEmitter->emitNode($args[1]);
-        $this->outputEmitter->emitStr('))', $loc);
-        return true;
-    }
-
-    /**
-     * `(empty? x)` on a tagged local — emit the native check
-     * specific to the tag, skipping the runtime cond chain.
-     */
-    private function tryEmitEmptyCheck(CallNode $node): bool
-    {
-        $fragment = TypedValueSpecialization::emptyCheckFragment($node);
-        if ($fragment === null) {
-            return false;
-        }
-
-        $loc = $node->getStartSourceLocation();
-        $parts = explode('%s', $fragment, 2);
-        assert(count($parts) === 2);
-
-        $this->outputEmitter->emitStr($parts[0], $loc);
-        $this->outputEmitter->emitNode($node->getArguments()[0]);
-        $this->outputEmitter->emitStr($parts[1], $loc);
-        return true;
-    }
-
-    /**
-     * `(name x)` / `(namespace x)` on a target tagged
-     * `\Phel\Lang\Keyword` or `\Phel\Lang\Symbol` — emit the direct
-     * method call, skipping the runtime cond chain.
-     */
-    private function tryEmitNamedAccessor(CallNode $node): bool
-    {
-        $method = TypedValueSpecialization::namedAccessorMethod($node);
-        if ($method === null) {
-            return false;
-        }
-
-        $loc = $node->getStartSourceLocation();
-        $this->outputEmitter->emitStr('(', $loc);
-        $this->outputEmitter->emitNode($node->getArguments()[0]);
-        $this->outputEmitter->emitStr('->' . $method . '())', $loc);
-        return true;
-    }
-
-    /**
-     * Splice the argument into the native predicate fragment from
-     * `TypePredicateSpecialization::typePredicateFragment`. The fragment uses
-     * `sprintf` placeholders so multi-instanceof predicates (`map?`,
-     * `vector?`, `seq?`) can reference the argument multiple times.
-     */
-    private function tryEmitTypePredicate(CallNode $node): bool
-    {
-        $fragment = TypePredicateSpecialization::typePredicateFragment($node);
-        if ($fragment === null) {
-            return false;
-        }
-
-        $loc = $node->getStartSourceLocation();
-
-        $argEmit = $this->outputEmitter->captureNodeAsExpression($node->getArguments()[0]);
-
-        $this->outputEmitter->emitStr(sprintf($fragment, $argEmit), $loc);
-        return true;
-    }
-
-    /**
-     * `(zero? x)` / `(pos? x)` / `(neg? x)` on an `int` / `float`
-     * tagged local — emit the native comparison directly.
-     */
-    private function tryEmitNumericPredicate(CallNode $node): bool
-    {
-        $name = TypePredicateSpecialization::isNumericPredicate($node);
-        if ($name === null) {
-            return false;
-        }
-
-        $loc = $node->getStartSourceLocation();
-        $this->outputEmitter->emitStr('(', $loc);
-        $this->outputEmitter->emitNode($node->getArguments()[0]);
-
-        $op = match ($name) {
-            'zero?' => ' === 0',
-            'pos?' => ' > 0',
-            'neg?' => ' < 0',
-            default => ' === 0',
-        };
-
-        $this->outputEmitter->emitStr($op . ')', $loc);
-        return true;
-    }
-
-    private function tryEmitTypedPhpArrayGet(CallNode $node): bool
-    {
-        if (!CallSpecialization::isTypedPhpArrayGet($node)) {
-            return false;
-        }
-
-        $args = $node->getArguments();
-        $loc = $node->getStartSourceLocation();
-        $this->outputEmitter->emitStr('(', $loc);
-        $this->outputEmitter->emitNode($args[0]);
-        $this->outputEmitter->emitStr('[', $loc);
-        $this->outputEmitter->emitNode($args[1]);
-        $this->outputEmitter->emitStr('] ?? ', $loc);
-        if (count($args) === 3) {
-            $this->outputEmitter->emitNode($args[2]);
-        } else {
-            $this->outputEmitter->emitStr('null', $loc);
-        }
-
-        $this->outputEmitter->emitStr(')', $loc);
-        return true;
-    }
-
-    /**
-     * `(not (= a b))` peephole over the typed-`=` specialiser:
-     * emits `($a !== $b)` directly, skipping both `phel.core/not`
-     * and the explicit `!(($a === $b))` wrapper.
-     *
-     * Eligibility lives on {@see NumericOperationSpecialization::notEqPeepholeInner()}.
-     */
-    private function tryEmitNotEqPeephole(CallNode $node): bool
-    {
-        $inner = NumericOperationSpecialization::notEqPeepholeInner($node);
-        if (!$inner instanceof CallNode) {
-            return false;
-        }
-
-        $args = $inner->getArguments();
-        $loc = $node->getStartSourceLocation();
-        $this->outputEmitter->emitStr('(', $loc);
-        $this->outputEmitter->emitNode($args[0]);
-        $this->outputEmitter->emitStr(' !== ', $loc);
-        $this->outputEmitter->emitNode($args[1]);
-        $this->outputEmitter->emitStr(')', $loc);
-        return true;
-    }
-
-    /**
-     * Specialise a chain of `(assoc m k v)` calls (or `(conj v x)`
-     * calls) on a typed persistent target — after thread-macro
-     * expansion these are nested `CallNode`s of the same op rooted at
-     * a `LocalVarNode`. The runtime path goes through one persistent
-     * `put` / `append` per chain step, each allocating a new persistent
-     * map / vector. The transient path opens one transient at the leaf
-     * target, mutates it once per chain step, and snapshots back to a
-     * persistent at the end — N-1 persistent intermediates collapse to
-     * one.
-     */
-    private function tryEmitAssocConjChain(CallNode $node): bool
-    {
-        $chain = AssocConjSpecialization::assocConjChain($node);
-        if ($chain === null) {
-            return false;
-        }
-
-        $loc = $node->getStartSourceLocation();
-        $this->outputEmitter->emitStr('((', $loc);
-        $this->outputEmitter->emitNode($chain['target']);
-        $this->outputEmitter->emitStr(')->asTransient()', $loc);
-
-        foreach ($chain['groups'] as $group) {
-            $this->outputEmitter->emitStr('->' . $chain['method'] . '(', $loc);
-            foreach ($group as $i => $arg) {
-                if ($i > 0) {
-                    $this->outputEmitter->emitStr(', ', $loc);
-                }
-
-                $this->outputEmitter->emitNode($arg);
-            }
-
-            $this->outputEmitter->emitStr(')', $loc);
-        }
-
-        $this->outputEmitter->emitStr('->persistent())', $loc);
-        return true;
-    }
-
-    /**
-     * Specialise `(assoc m k v)` / `(assoc v i x)` / `(conj v x)` /
-     * `(dissoc m k)` to a direct persistent-collection method when the
-     * target tag is known. Skips variadic forms.
-     */
-    private function tryEmitTypedAssocConjDissoc(CallNode $node): bool
-    {
-        $method = AssocConjSpecialization::typedAssocConjDissocMethod($node);
-        if ($method === null) {
-            return false;
-        }
-
-        $args = $node->getArguments();
-        $loc = $node->getStartSourceLocation();
-        $this->outputEmitter->emitStr('(', $loc);
-        $this->outputEmitter->emitNode($args[0]);
-        $this->outputEmitter->emitStr('->' . $method . '(', $loc);
-
-        $rest = array_slice($args, 1);
-        $count = count($rest);
-        foreach ($rest as $i => $arg) {
-            $this->outputEmitter->emitNode($arg);
-            if ($i < $count - 1) {
-                $this->outputEmitter->emitStr(', ', $loc);
-            }
-        }
-
-        $this->outputEmitter->emitStr('))', $loc);
-        return true;
-    }
-
-    /**
-     * Specialise `(first s)` / `(rest s)` to a direct method call on
-     * the tagged seq target. The runtime `phel.core/first` and
-     * `phel.core/rest` bodies walk cond chains over nil / string /
-     * php-array / set / map / seq; for a known seq tag every branch
-     * collapses to `$s->first()` / `$s->rest()`.
-     */
-    private function tryEmitTypedSeqAccessor(CallNode $node): bool
-    {
-        $method = TypedCollectionMethodSpecialization::typedSeqMethodName($node);
-        if ($method === null) {
-            return false;
-        }
-
-        $args = $node->getArguments();
-        $loc = $node->getStartSourceLocation();
-        $this->outputEmitter->emitStr('(', $loc);
-        $this->outputEmitter->emitNode($args[0]);
-        $this->outputEmitter->emitStr('->' . $method . '())', $loc);
-        return true;
-    }
-
-    /**
-     * Specialise `(nth v i)` / `(count v)` to a direct method call on
-     * the tagged `PersistentVectorInterface` target. The runtime
-     * `phel.core/nth` body walks a `cond` over set / seq / vector /
-     * map / php-array; for a typed vector every branch collapses to
-     * a single method call.
-     */
-    private function tryEmitTypedVectorAccessor(CallNode $node): bool
-    {
-        $spec = TypedCollectionMethodSpecialization::typedVectorMethodCall($node);
-        if ($spec === null) {
-            return false;
-        }
-
-        $args = $node->getArguments();
-        $loc = $node->getStartSourceLocation();
-        $this->outputEmitter->emitStr('(', $loc);
-        $this->outputEmitter->emitNode($args[0]);
-        $this->outputEmitter->emitStr('->' . $spec['method'] . '(', $loc);
-
-        $argCount = count($spec['args']);
-        foreach ($spec['args'] as $i => $argIndex) {
-            $this->outputEmitter->emitNode($args[$argIndex]);
-            if ($i < $argCount - 1) {
-                $this->outputEmitter->emitStr(', ', $loc);
-            }
-        }
-
-        $this->outputEmitter->emitStr('))', $loc);
-        return true;
     }
 
     private function emitPhpFunctionName(PhpVarNode $fnNode): void
